Revert "rename the Import directory to import, which is consistent with other packages"
This reverts commit a274ce52ed2c2710f442e5db2df752c1459670c2.
